backend/requirements.txt
fastapi>=0.100.0,<0.101.0
uvicorn[standard]>=0.23.0,<0.24.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
qdrant-client>=1.9.0,<2.0.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
httpx>=0.25.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
trafilatura>=1.6.0
langchain-text-splitters>=0.0.1
requests>=2.31.0
transformers>=4.36.0
diffusers>=0.25.0
accelerate>=0.25.0
Pillow>=10.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0